Add support for JupyterHub, utilizing JupyterHub API token
Fixes https://github.com/msm1089/ipynbname/issues/8

This utilizes the `JUPYTERHUB_API_TOKEN` environment variable as a
fallback to get the token when running notebooks in a JupyterHub
context, since the token is *not* contained in the server info files.

As demonstrated similarly in this comment:
https://github.com/jupyterhub/jupyterhub/issues/1718#issuecomment-377036175

This simple change worked without issue for me when tested on
JupyterHub.